+++ src/re_com/util.cljs
@@ -25,7 +25,7 @@
 (defn deref-or-value
   "Takes a value or an atom
   If it's a value, returns it
-  If it's a Reagent object that supports IDeref, returns the value inside it by derefing
+  If it's an object that supports IDeref, returns the value inside it by derefing
   "
   [val-or-atom]
   (if (satisfies? IDeref val-or-atom)